requirements.txt
python-wakeonlan
requests
pyyaml
psutil
